Triggers nuevos, Procedures con transacciones y checks - - hay que cambiar varios triggers por checks

+++ hospital-db/04_tables/03_medico.sql
@@ -6,7 +6,7 @@ GO
 CREATE TABLE Internacion.Medico (
     matricula matricula_type PRIMARY KEY,
     cuil VARCHAR(13) NOT NULL,
-    cantMaxGuardia INT NOT NULL,
+    cantMaxGuardia TINYINT NOT NULL,
     fechaIngreso DATE NOT NULL,
     foto imagen_type NOT NULL,
     dni dni_type NOT NULL,
