+++ hospital-db/04_tables/07_habitacion.sql
@@ -3,7 +3,7 @@ GO
 
 CREATE TABLE Internacion.Habitacion (
     idHabitacion INT IDENTITY(1,1) PRIMARY KEY NOT NULL,
-    piso INT NOT NULL,
+    piso TINYINT NOT NULL,
     orientacion NVARCHAR(20) NOT NULL, -- CAMBIAR
     idSector INT NOT NULL,
 );
