+++ hospital-db/04_tables/08_cama.sql
@@ -3,7 +3,7 @@ GO
 
 CREATE TABLE Internacion.Cama (
     idHabitacion INT NOT NULL,
-    nroCama INT NOT NULL,
+    nroCama TINYINT NOT NULL,
     ocupada BIT NOT NULL,
     PRIMARY KEY (idHabitacion, nroCama),
 );
