+++ hospital-db/04_tables/10_ubicada_en.sql
@@ -3,7 +3,7 @@ GO
 
 CREATE TABLE Internacion.UbicadaEn (
     idHabitacion INT NOT NULL,
-    nroCama INT NOT NULL,
+    nroCama TINYINT NOT NULL,
     fechaYHora DATETIME NOT NULL,
     idInternacion INT NOT NULL,
     PRIMARY KEY (idHabitacion, nroCama, fechaYHora),
